requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.6
mysql-connector-python==8.3.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
